requirements.txt
Pillow==8.3.2
tflite-runtime==2.6.0